Update libsa to follow std c lib prototypes
git-svn-id: http://forge.voodooprojects.org/svn/chameleon/trunk@865 b5af796a-00a8-11df-919b-ffff7a100b7d

+++ i386/libsa/string.c
@@ -119,7 +119,7 @@ int strlen(const char * s)
 
 /* NOTE: Moved from ntfs.c */
 int
-memcmp(const void *p1, const void *p2, int len)
+memcmp(const void *p1, const void *p2, size_t len)
 {
     while (len--) {
         if (*(const char*)(p1++) != *(const char*)(p2++))
